data in ms to sec

+++ app/src/main/java/com/aafiyahtech/ventilator/models/Group_3_B.kt
@@ -5,7 +5,7 @@ import com.google.gson.annotations.SerializedName
 data class Group_3_B(
     @SerializedName("PlatePressure") val platePressure: Float,
     @SerializedName("Acceleration") val acceleration: Float,
-    @SerializedName("InspiratoryONDelay") val inspiratoryONDelay: Int,
+    @SerializedName("InspiratoryONDelay") val inspiratoryONDelay: Float,
     @SerializedName("tPause") val tPause: Int,
     @SerializedName("dummy1") val dummy1: Float
